test,win: split addon tests
