+++ test/addons/esm-export-default/binding-export-default.cc
