+++ test/addons/esm-export-primitive/binding-export-primitive.cc
