+++ test/addons/hello-world-2/binding.cc
